refactor(auth): isolar fluxo de identidade e especializar controller de usuário
