+++ src/main/resources/templates/user/dashboard.html
@@ -7,7 +7,7 @@
     <main class="container" style="margin-top: 4rem;">
         <div class="animate-fade">
             <h1 style="font-family: 'Outfit', sans-serif; font-size: 2.5rem; color: var(--cor-primary-dark);">
-                Olá, <span th:text="${user.name}">Usuário</span>!
+                Olá, <span th:text="${currentUser.name}">Usuário</span>!
             </h1>
             <p style="color: var(--cor-text-secondary); margin-bottom: 3rem;">Bem-vindo ao seu painel de estudos em Libras.</p>
