+++ src/main/resources/templates/user/profile.html
@@ -10,14 +10,14 @@
                 👤
             </div>
             
-            <h1 th:text="${user.name}" style="font-family: 'Outfit', sans-serif;">Nome do Usuário</h1>
-            <p th:text="${user.email}" style="color: var(--cor-text-secondary); margin-bottom: 3rem;">email@exemplo.com</p>
+            <h1 th:text="${currentUser.name}" style="font-family: 'Outfit', sans-serif;">Nome do Usuário</h1>
+            <p th:text="${currentUser.email}" style="color: var(--cor-text-secondary); margin-bottom: 3rem;">email@exemplo.com</p>
             
             <hr style="border: none; border-top: 1px solid var(--cor-border); margin-bottom: 2rem;">
             
             <div style="text-align: left; margin-bottom: 3rem;">
-                <p><strong>Nível de Acesso:</strong> <span class="categoria-tag" th:text="${user.role}">USER</span></p>
-                <p style="margin-top: 1rem;"><strong>Desde:</strong> <span th:text="${#temporals.format(user.createdAt, 'dd/MM/yyyy')}">01/01/2026</span></p>
+                <p><strong>Nível de Acesso:</strong> <span class="categoria-tag" th:text="${currentUser.role}">USER</span></p>
+                <p style="margin-top: 1rem;"><strong>Desde:</strong> <span th:text="${#temporals.format(currentUser.createdAt, 'dd/MM/yyyy')}">01/01/2026</span></p>
             </div>
             
             <div style="display: flex; flex-direction: column; gap: 1rem;">
